--- common/testing.py
@@ -1,8 +1,0 @@
-import os
-from nose.tools import nottest
-
-def phone_only(x):
-  if os.path.isfile("/init.qcom.rc"):
-    return x
-  else:
-    return nottest(x)
