+++ selfdrive/test/test_openpilot.py
@@ -2,77 +2,22 @@
 import os
 os.environ['FAKEUPLOAD'] = "1"
 
-from common.apk import update_apks, start_offroad, pm_apply_packages, android_packages
 from common.params import Params
 from common.realtime import sec_since_boot
-from common.testing import phone_only
-from selfdrive.manager import manager_init, manager_prepare
-from selfdrive.manager import start_managed_process, kill_managed_process, get_running
-from selfdrive.manager import start_daemon_process
-from functools import wraps
+from selfdrive.manager import manager_init, manager_prepare, start_daemon_process
+from selfdrive.test.helpers import phone_only, with_processes
 import json
 import requests
 import signal
 import subprocess
 import time
 
-DID_INIT = False
 
 # must run first
 @phone_only
 def test_manager_prepare():
-  global DID_INIT
   manager_init()
   manager_prepare()
-  DID_INIT = True
-
-def with_processes(processes):
-  def wrapper(func):
-    @wraps(func)
-    def wrap():
-      if not DID_INIT:
-        test_manager_prepare()
-
-      # start and assert started
-      [start_managed_process(p) for p in processes]
-      assert all(get_running()[name].exitcode is None for name in processes)
-
-      # call the function
-      try:
-        func()
-        # assert processes are still started
-        assert all(get_running()[name].exitcode is None for name in processes)
-      finally:
-        # kill and assert all stopped
-        [kill_managed_process(p) for p in processes]
-        assert len(get_running()) == 0
-    return wrap
-  return wrapper
-
-def with_apks():
-  def wrapper(func):
-    @wraps(func)
-    def wrap():
-      if not DID_INIT:
-        test_manager_prepare()
-
-      update_apks()
-      pm_apply_packages('enable')
-      start_offroad()
-
-      func()
-
-      try:
-        for package in android_packages:
-          apk_is_running = (subprocess.call(["pidof", package]) == 0)
-          assert apk_is_running, package
-      finally:
-        pm_apply_packages('disable')
-        for package in android_packages:
-          apk_is_not_running = (subprocess.call(["pidof", package]) == 1)
-          assert apk_is_not_running, package
-    return wrap
-  return wrapper
 
 @phone_only
 @with_processes(['loggerd', 'logmessaged', 'tombstoned', 'proclogd', 'logcatd'])
